requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.0
channels~=3.0.5
channels-redis==3.4.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.1
django-4-jet==1.0.9
django-activity-stream==1.4.1
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-hstore==1.4.2
django-jsonfield-backport==1.0.5
django-model-utils==4.2.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
frozenlist==1.3.0
geoip2==4.5.0
google-api-python-client==1.4.1
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
influxdb==5.3.1
influxdb-client==1.25.0
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
kombu==5.2.3
MarkupSafe==2.0.1
maxminddb==2.2.0
mqttasgi==1.1.2
msgpack==1.0.4
multidict==6.0.2
numpy==1.22.1
oauth2client==4.1.3
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-memcached==1.59
pytz==2021.3
redis==4.1.4
requests==2.27.1
rsa==4.9
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
Rx==3.2.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.7.2
zope.interface==5.4.0

django-notifications-hq~=1.7.0
django-softdelete~=0.9.4